:art: Cleaned the code a bit
Removed try except block for checking if the config exists, replaced by `os.path.isfile`. Also added checks for when to write to the config file. This is because previously we would keep writing the same token to the same config each run. Now it only writes the token to the config if the config doesn't exist yet. Also did some minor things like change around comments.

+++ index.py
@@ -2,10 +2,11 @@ import requests
 import json
 import inspect
 
+from os import path
 from colorama import Fore, Style
 from discord import app_commands, Intents, Client, Interaction
 
-# ASCII logo, uses Colorama for coloring the logo. 
+# ASCII logo, uses Colorama for coloring the logo.
 logo = f"""
 {Fore.LIGHTBLUE_EX}       {Fore.GREEN}cclloooooooooooooo.
 {Fore.LIGHTBLUE_EX},;;;:{Fore.GREEN}oooooooooooooooooooooo.
@@ -34,19 +35,17 @@ print(logo + inspect.cleandoc(f"""
     You may close it after the bot has been invited and the command has been ran{Style.RESET_ALL}
 """))
 
-# Try except block is useful for when you'd like to capture errors
-try:
+# Checks if a config file has been made
+# If so, load said config, else load no config
+if path.isfile("config.json"):
     with open("config.json") as f:
         config = json.load(f)
-except (FileNotFoundError, json.JSONDecodeError):
-    # You can in theory also do "except:" or "except Exception:", but it is not recommended
-    # unless you want to suppress all errors
+else:
     config = {}
 
 
 while True:
-    # Attempts to pull a token from config
-    # If no token is stored the token variable stores value None
+    # If no token is stored in "config" the value defaults to None
     token = config.get("token", None)
     if token:
         print(f"\n--- Detected token in {Fore.GREEN}./config.json{Fore.RESET} (saved from a previous run). Using stored token. ---\n")
@@ -56,7 +55,7 @@ while True:
 
     # Validates if the token you provided was correct or not
     # There is also another one called aiohttp.ClientSession() which is asynchronous
-    # However for such simplicity, it is not worth playing around with async 
+    # However for such simplicity, it is not worth playing around with async
     # and await keywords outside of the event loop
     try:
         data = requests.get("https://discord.com/api/v10/users/@me", headers={
@@ -64,13 +63,13 @@ while True:
         }).json()
     except requests.exceptions.RequestException as e:
         if e.__class__ == requests.exceptions.ConnectionError:
-            exit("ConnectionResetError: Discord is commonly blocked on public networks, please make sure discord.com is reachable!")
-        
+            exit(f"{Fore.RED}ConnectionError{Fore.RESET}: Discord is commonly blocked on public networks, please make sure discord.com is reachable!")
+
         elif e.__class__ == requests.exceptions.Timeout:
-            exit("Timeout: Connection to Discord's API has timed out (possibly being rate limited?)")
-        
+            exit(f"{Fore.RED}Timeout{Fore.RESET}: Connection to Discord's API has timed out (possibly being rate limited?)")
+
         # Tells python to quit, along with printing some info on the error that occured
-        raise exit(f"Unknown error has occurred! Additional info:\n{e}")
+        exit(f"Unknown error has occurred! Additional info:\n{e}")
 
     # If the token is correct, it will continue the code
     if data.get("id", None):
@@ -81,18 +80,19 @@ while True:
     print(f"\nSeems like you entered an {Fore.RED}invalid token{Fore.RESET}. Please enter a valid token (see Github repo for help).")
 
     # Resets the config so that it doesn't use the previous token again
-    config = {}
+    config.clear()
 
 
-# This is used to save the token for the next time you run the bot
-with open("config.json", "w") as f:
-    # Check if 'token' key exists in the config.json file
-    config["token"] = token
+# Save token into config file for next run if config.json doesn't exsist already
+if not path.isfile("config.json"):
+    with open("config.json", "w") as f:
+        # Create new key 'token' storing our bot's token
+        config["token"] = token
 
-    # This dumps our working setting to the config.json file
-    # Indent is used to make the file look nice and clean
-    # If you don't want to indent, you can remove the indent=2 from code
-    json.dump(config, f, indent=2)
+        # This dumps our working setting to the config.json file
+        # Indent is used to make the file look nice and clean
+        # If you don't want to indent, you can remove the indent=2 from code
+        json.dump(config, f, indent=2)
 
 
 class FunnyBadge(Client):
@@ -102,7 +102,7 @@ class FunnyBadge(Client):
 
     async def setup_hook(self) -> None:
         """ This is called when the bot boots, to setup the global commands """
-        await self.tree.sync(guild=None)
+        await self.tree.sync()
 
 
 # Variable to store the bot class and interact with it
